+++ lib/SILGen/SILGen.cpp
@@ -2,7 +2,7 @@
 //
 // This source file is part of the Swift.org open source project
 //
-// Copyright (c) 2014 - 2017 Apple Inc. and the Swift project authors
+// Copyright (c) 2014 - 2020 Apple Inc. and the Swift project authors
 // Licensed under Apache License v2.0 with Runtime Library Exception
 //
 // See https://swift.org/LICENSE.txt for license information
@@ -1946,6 +1946,12 @@ ASTLoweringRequest::evaluate(Evaluator &evaluator,
 
   auto silMod = SILModule::createEmptyModule(desc.context, desc.conv,
                                              desc.opts);
+
+  // If all function bodies are being skipped there's no reason to do any
+  // SIL generation.
+  if (desc.opts.SkipFunctionBodies == FunctionBodySkipping::All)
+    return silMod;
+
   SILGenModuleRAII scope(*silMod);
 
   // Emit a specific set of SILDeclRefs if needed.
