+++ lib/Sema/TypeCheckDeclPrimary.cpp
@@ -2,7 +2,7 @@
 //
 // This source file is part of the Swift.org open source project
 //
-// Copyright (c) 2014 - 2018 Apple Inc. and the Swift project authors
+// Copyright (c) 2014 - 2020 Apple Inc. and the Swift project authors
 // Licensed under Apache License v2.0 with Runtime Library Exception
 //
 // See https://swift.org/LICENSE.txt for license information
@@ -2324,8 +2324,9 @@ public:
 
 
   bool shouldSkipBodyTypechecking(const AbstractFunctionDecl *AFD) {
-    // Make sure we're in the mode that's skipping function bodies.
-    if (!getASTContext().TypeCheckerOpts.SkipNonInlinableFunctionBodies)
+    // Make sure we're in a mode that's skipping function bodies.
+    if (getASTContext().TypeCheckerOpts.SkipFunctionBodies ==
+        FunctionBodySkipping::None)
       return false;
 
     // Make sure there even _is_ a body that we can skip.
@@ -2339,11 +2340,13 @@ public:
         return false;
     }
 
-    // If we're gonna serialize the body, we can't skip it.
-    if (AFD->getResilienceExpansion() == ResilienceExpansion::Minimal)
-      return false;
+    // Skipping all bodies won't serialize anything, so can skip regardless
+    if (getASTContext().TypeCheckerOpts.SkipFunctionBodies ==
+        FunctionBodySkipping::All)
+      return true;
 
-    return true;
+    // Only skip functions where their body won't be serialized
+    return AFD->getResilienceExpansion() != ResilienceExpansion::Minimal;
   }
 
   void visitFuncDecl(FuncDecl *FD) {
